refactor userprofile component out to another file, also extracted menus
